// src/app/core/models/liquidacionTerap.ts
export class LiquidacionTerapeuta {
    currentDate?: string;
    desdeFechaLiquidado?: string;
    desdeHoraLiquidado?: string;
    encargada?: string;
    hastaFechaLiquidado?: string;
    hastaHoraLiquidado?: string;
    id?: number;
    idTerapeuta?: number;
    importe?: number;
    terapeuta?: string;
    tratamiento?: number;
  }
  